requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==1.4.51
pymysql==1.1.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.0
feast>=0.32.0
joblib==1.3.2
requests==2.31.0
gradio>=4.0.0
python-multipart==0.0.6
alembic==1.12.1